chore: get changes on individual configs
