+++ src/api/services/configs.js
@@ -11,6 +11,9 @@ export const getAllChanges = async () =>
 export const getConfig = async (id) =>
   resolve(ConfigDB.get(`/config_item?id=eq.${id}`));
 
+export const getConfigChange = async (id) =>
+  resolve(ConfigDB.get(`/config_change?config_id=eq.${id}`));
+
 export const createConfigItem = async (type, params) =>
   resolve(
     ConfigDB.post(`/config_item`, {
